pyproject.toml
[tool.poetry]
name = "superkeet"
version = "0.1.0"
description = "FOSS voice-to-text application for macOS"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
parakeet-mlx = "^0.3.0"
PySide6 = "^6.6.0"
sounddevice = "^0.4.6"
pynput = "^1.7.6"
pyperclip = "^1.8.2"
numpy = "==2.2.5"
pyyaml = "^6.0"
pyqtgraph = "^0.13.7"
soundfile = "^0.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true